@@ -58,26 +65,26 @@
     "watch": "npm run watch:dev",
     "webdriver-manager": "webdriver-manager",
     "webdriver:start": "npm run webdriver-manager start",
-    "webdriver:update": "npm run webdriver-manager update",
+    "webdriver:update": "webdriver-manager update",
     "webpack-dev-server": "webpack-dev-server",
     "webpack": "webpack"
   },
   "dependencies": {
-    "@angular/common": "2.2.1",
-    "@angular/compiler": "2.2.1",
-    "@angular/core": "2.2.1",
-    "@angular/forms": "2.2.1",
-    "@angular/http": "2.2.1",
-    "@angular/platform-browser": "2.2.1",
-    "@angular/platform-browser-dynamic": "2.2.1",
-    "@angular/platform-server": "2.2.1",
-    "@angular/router": "3.2.1",
+    "@angular/common": "~2.4.3",
+    "@angular/compiler": "~2.4.3",
+    "@angular/core": "~2.4.3",
+    "@angular/forms": "~2.4.3",
+    "@angular/http": "~2.4.3",
+    "@angular/platform-browser": "~2.4.3",
+    "@angular/platform-browser-dynamic": "~2.4.3",
+    "@angular/platform-server": "~2.4.3",
+    "@angular/router": "~3.4.3",
     "@angularclass/conventions-loader": "^1.0.2",
     "@angularclass/hmr": "~1.2.2",
     "@angularclass/hmr-loader": "~3.0.2",
     "assets-webpack-plugin": "^3.4.0",
     "brace": "^0.8.0",
-    "core-js": "^2.4.0",
+    "core-js": "^2.4.1",
     "d3": "^4.4.0",
     "d3-bboxCollide": "^1.0.3",
     "firebase": "~2.4.2",
@@ -163,18 +181,16 @@
     "susy": "^2.2.12",
     "to-string-loader": "^1.1.4",
     "ts-helpers": "1.1.2",
-    "ts-node": "^1.7.0",
-    "tslint": "~3.15.1",
-    "tslint-loader": "^2.1.5",
-    "typedoc": "^0.5.0",
-    "typescript": "2.1.1",
+    "ts-node": "^2.0.0",
+    "tslint": "~4.4.0",
+    "typedoc": "^0.5.3",
+    "typescript": "~2.1.5",
     "url-loader": "^0.5.7",
-    "v8-lazy-parse-webpack-plugin": "^0.3.0",
-    "webpack": "2.1.0-beta.27",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-dev-server": "2.1.0-beta.11",
-    "webpack-md5-hash": "^0.0.5",
-    "webpack-merge": "~1.0.1"
+    "webpack": "2.2.0",
+    "webpack-dev-middleware": "^1.9.0",
+    "webpack-dev-server": "2.2.0",
+    "webpack-dll-bundles-plugin": "^1.0.0-beta.5",
+    "webpack-merge": "~2.4.0"
   },
   "engines": {
     "node": ">= 4.2.1",
