@@ -3,6 +3,8 @@
  */
 var path = require('path');
 
+const EVENT = process.env.npm_lifecycle_event || '';
+
 // Helper functions
 var ROOT = path.resolve(__dirname, '..');
 
@@ -10,15 +12,17 @@ function hasProcessFlag(flag) {
   return process.argv.join('').indexOf(flag) > -1;
 }
 
+function hasNpmFlag(flag) {
+  return EVENT.includes(flag);
+}
+
 function isWebpackDevServer() {
   return process.argv[1] && !! (/webpack-dev-server/.exec(process.argv[1]));
 }
 
-function root(args) {
-  args = Array.prototype.slice.call(arguments, 0);
-  return path.join.apply(path, [ROOT].concat(args));
-}
+var root = path.join.bind(path, ROOT);
 
 exports.hasProcessFlag = hasProcessFlag;
+exports.hasNpmFlag = hasNpmFlag;
 exports.isWebpackDevServer = isWebpackDevServer;
 exports.root = root;
