@@ -1,10 +1,12 @@
 const helpers = require('./helpers');
 const webpackMerge = require('webpack-merge'); // used to merge webpack configs
+const webpackMergeDll = webpackMerge.strategy({plugins: 'replace'});
 const commonConfig = require('./webpack.common.js'); // the settings that are common to prod and dev
 
 /**
  * Webpack Plugins
  */
+const AddAssetHtmlPlugin = require('add-asset-html-webpack-plugin');
 const DefinePlugin = require('webpack/lib/DefinePlugin');
 const NamedModulesPlugin = require('webpack/lib/NamedModulesPlugin');
 const LoaderOptionsPlugin = require('webpack/lib/LoaderOptionsPlugin');
@@ -22,6 +24,10 @@ const METADATA = webpackMerge(commonConfig({env: ENV}).metadata, {
   ENV: ENV,
   HMR: HMR
 });
+
+const autoprefixer = require('autoprefixer');
+
+const DllBundlesPlugin = require('webpack-dll-bundles-plugin').DllBundlesPlugin;
 
 /**
  * Webpack configuration
@@ -67,7 +73,7 @@ module.exports = function (options) {
        *
        * See: http://webpack.github.io/docs/configuration.html#output-sourcemapfilename
        */
-      sourceMapFilename: '[name].map',
+      sourceMapFilename: '[file].map',
 
       /** The filename of non-entry chunks as relative path
        * inside the output.path directory.
@@ -78,6 +84,59 @@ module.exports = function (options) {
 
       library: 'ac_[name]',
       libraryTarget: 'var',
+    },
+
+    module: {
+
+      rules: [
+
+        /*
+         * css loader support for *.css files (styles directory only)
+         * Loads external css styles into the DOM, supports HMR
+         *
+         */
+        {
+          test: /\.css$/,
+          use: ['style-loader', 'css-loader'],
+        },
+
+        /*
+         * sass loader support for *.scss files (styles directory only)
+         * Loads external sass styles into the DOM, supports HMR
+         *
+         */
+        {
+          test: /\.scss$/,
+          use: ['style-loader', 'css-loader', 'sass-loader'],
+          include: [helpers.root('src', 'styles')]
+        },
+
+        {
+          enforce: 'pre',
+          test: /\.sass$/,
+          include: [helpers.root('src', 'styles')],
+          use: 'import-glob-loader'
+        },
+        {
+          test: /\.sass$/,
+          include: [helpers.root('src', 'styles')],
+          use: [
+            'style-loader',
+            'css-loader?importLoaders=1&sourceMap',
+            {
+              loader: 'postcss-loader',
+              options: {
+                sourceMap: 'inline',
+                parser: 'postcss-scss',
+                plugins: () => [autoprefixer],
+              }
+            },
+            'sass-loader?sourceMap',
+          ]
+        },
+
+      ]
+
     },
 
     plugins: [
@@ -102,6 +161,84 @@ module.exports = function (options) {
         }
       }),
 
+      new DllBundlesPlugin({
+        bundles: {
+          polyfills: [
+            'core-js',
+            {
+              name: 'zone.js',
+              path: 'zone.js/dist/zone.js'
+            },
+            {
+              name: 'zone.js',
+              path: 'zone.js/dist/long-stack-trace-zone.js'
+            },
+            'ts-helpers',
+          ],
+          vendor: [
+            '@angular/platform-browser',
+            '@angular/platform-browser-dynamic',
+            '@angular/core',
+            '@angular/common',
+            '@angular/forms',
+            '@angular/http',
+            '@angular/router',
+            '@angularclass/hmr',
+            'rxjs',
+
+            // Ace editor
+            'brace',
+            {
+              name: 'brace',
+              path: 'brace/mode/javascript.js',
+            },
+            // {
+            //   name: 'brace',
+            //   path: 'brace/theme/monokai.js',
+            // },
+
+            // The rest
+            'd3',
+            'd3-bboxCollide',
+            'firebase',
+            'hammer-timejs',
+            'intl',
+            {
+              name: 'intl',
+              path: 'intl/locale-data/jsonp/en.js',
+            },
+            'jquery',
+            'lodash',
+            'lunr',
+            'mousetrap',
+            'mousetrap-global-bind',
+            'sentiment',
+            'toastr',
+
+            // This didn't work cause this plugin needs npm packages (for name and version to know how to cache them). See my comments here <https://github.com/shlomiassaf/webpack-dll-bundles-plugin/issues/8>. For now this library will just have to not be in the vendor bundle
+            // process.cwd() + '/src/app/vendor/darsain-tooltips.js',
+          ]
+        },
+        dllDir: helpers.root('dll'),
+        webpackConfig: webpackMergeDll(commonConfig({env: ENV}), {
+          devtool: 'cheap-module-source-map',
+          plugins: []
+        })
+      }),
+
+      /**
+       * Plugin: AddAssetHtmlPlugin
+       * Description: Adds the given JS or CSS file to the files
+       * Webpack knows about, and put it into the list of assets
+       * html-webpack-plugin injects into the generated html.
+       *
+       * See: https://github.com/SimenB/add-asset-html-webpack-plugin
+       */
+      new AddAssetHtmlPlugin([
+        { filepath: helpers.root(`dll/${DllBundlesPlugin.resolveFile('polyfills')}`) },
+        { filepath: helpers.root(`dll/${DllBundlesPlugin.resolveFile('vendor')}`) }
+      ]),
+
       /**
        * Plugin: NamedModulesPlugin (experimental)
        * Description: Uses file names as module name.
@@ -118,7 +255,11 @@ module.exports = function (options) {
       new LoaderOptionsPlugin({
         debug: true,
         options: {
-
+          // These options via <https://github.com/shakacode/bootstrap-loader/issues/191>
+          context: helpers.root(),
+          output: {
+            path: helpers.root('dist')
+          },
         }
       }),
 
