@@ -18,9 +18,37 @@ export function main(): Promise<any> {
   return platformBrowserDynamic()
     .bootstrapModule(AppModule)
     .then(decorateModuleRef)
-    .catch(err => console.error(err));
+    .catch((err) => console.error(err));
 }
 
 // needed for hmr
 // in prod this is replace for document ready
 bootloader(main);
+
+
+/*
+ * Vendor dependencies
+ *
+ * To pack these in `vendor` bundle they have to go in `webpack.dev.js` too. See `src/app/vendor/README.md`.
+ */
+
+// Ace editor
+import 'brace';
+import 'brace/mode/javascript';
+// import 'brace/theme/monokai';
+
+// The rest
+import 'd3';
+import 'd3-bboxCollide';
+import 'firebase';
+import 'hammer-timejs';
+import 'intl';
+import 'intl/locale-data/jsonp/en';
+import 'jquery';
+import 'lodash';
+import 'lunr';
+import 'mousetrap';
+import 'mousetrap-global-bind';
+import 'sentiment';
+import 'toastr';
+import 'app/vendor/darsain-tooltips';
