@@ -4,7 +4,6 @@
 
 // import 'ie-shim'; // Internet Explorer 9 support
 
-
 // import 'core-js/es6';
 // Added parts of es6 which are necessary for your project or your browser support requirements.
 import 'core-js/es6/symbol';
@@ -36,12 +35,12 @@ import 'ts-helpers';
 if ('production' === ENV) {
   // Production
 
-
 } else {
-  // Development
 
+  // Development
   Error.stackTraceLimit = Infinity;
 
+  /* tslint:disable no-var-requires */
   require('zone.js/dist/long-stack-trace-zone');
 
 }
